fabric-ci-cd/pyproject.toml
[tool.poetry]
name = "fabric-ci-cd"
version = "0.1.0"
description = ""
authors = ["Austin M. Reppert <austinmreppert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fabricclientaio = {git = "https://github.com/AustinMReppert/fabric-client-aio"}
azure-identity = "^1.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
